Updated ASCII_thunder_ball.py to remove flikering
Instead of clearing the console to write the new frame, which is super slow, we move the command cursor to the origin (0,0) and write over the old frame.
We use the kernel32's handler to allow the console to use ANSI escape sequences.

Thanks to Paul Ami Jeanbourquin for the tip of "moving the cursor". It made my day brother!